Modified namespace and naming of files.
