--- src/Leaves/Users/Edit/UserEditModel.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Rhubarb\Scaffolds\Saas\Tenant\Leaves\Users\Edit;
-
-use \Rhubarb\Leaf\Crud\Leaves\CrudModel;
-
-class UserEditModel extends CrudModel
-{
-    public $roleId;
-}
