+++ src/Leaves/Users/UsersItemModel.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Rhubarb\Scaffolds\Saas\Tenant\Leaves\Users;
+
+use \Rhubarb\Leaf\Crud\Leaves\CrudModel;
+
+class UsersItemModel extends CrudModel
+{
+    public $roleId;
+}
